requirements.txt
streamlit==1.28.1
psycopg2-binary==2.9.7
pandas==2.1.1
python-dotenv==1.0.0
kafka-python
requests
weaviate-client